This is a draft of an idea I want feedback/thoughts/suggestions on.

I was working on a different issue, and I was having problems getting stable timing results.  I ran llama-bench, and was getting pretty different results with each run.  See [before.txt](https://github.com/ggerganov/llama.cpp/files/15123143/before.txt)   PP timings varied between 21.46 to 32.42 t/s.   TG was between 11.59 to 12.86 t/s.

So I looked at how the threading works in ``ggml_compute_forward_mul_mat``.  Since work is pre-dispatched to the thread, each thread will always do the same amount of work, and not adapt if one thread is stalled by the OS, that'll delay the whole operation. 
 So what I did was break it into 16*number_of_threads slices, and had it work through the slices.  I pulled all the guts of the operation out into it's own function, and used the existing logic that divides up into 16x more chunks.  Each thread then grabs an ID of what the next chunk of work to do.  This gave a more stable and faster timing results.  
[after.txt](https://github.com/ggerganov/llama.cpp/files/15123349/after.txt)     PP between 20.83-35.12 t/s  and TG between 12.82-13.47 t/s.

So the range of the average went from 1.27 to 0.65, along with the speed going up.   Do you have a better metric I should try for measuring this?  Maybe I need to do llama-bench with -r 100 or something silly and let it run overnight, and compare the two results?  I'm trying to improve my testing process so I'm more sure about the results.

What are your thoughts on the general idea of changing like this?  I don't like the global I had to add, but I didn't want to add it to ``ggml_compute_params`` because I'd have to add ``atomic.h`` to the header... but maybe that's okay?  I'll do more tests, and test it on Linux before posting this as a PR, plus I wanna fix that global, get better timing comparisons, etc, etc.