@Jeximo 2024-04-26
```
uname -a
Linux localhost 4.14.190-23725627-abG975WVLS8IWD1 #2 SMP PREEMPT Mon Apr 10 18:16:39 KST 2023 aarch64 Android
```
I'm not sure how likely this is to be merged, but it's faster:

**master**:
```
llama_print_timings:        load time =   10137.45 ms
llama_print_timings:      sample time =      30.83 ms /   159 runs   (    0.19 ms per token,  5156.98 tokens per second)
llama_print_timings: prompt eval time =   41829.56 ms /    59 tokens (  708.98 ms per token,     1.41 tokens per second)
llama_print_timings:        eval time =   59630.39 ms /   158 runs   (  377.41 ms per token,     2.65 tokens per second)
llama_print_timings:       total time =  106957.99 ms /   217 tokens
```

**pr**:
```
llama_print_timings:        load time =    9489.74 ms                                              
llama_print_timings:      sample time =      29.13 ms /   159 runs   (    0.18 ms per token,  5458.48 tokens per second)
llama_print_timings: prompt eval time =   18258.65 ms /    59 tokens (  309.47 ms per token,     3.23 tokens per second)
llama_print_timings:        eval time =   55050.81 ms /   158 runs   (  348.42 ms per token,     2.87 tokens per second)
llama_print_timings:       total time =   75499.96 ms /   217 tokens
```

Greater than 2x `prompt eval time` with 7b IQ_4XS on CPU