@p-e-w 2024-04-26
Judging by @Jeximo's numbers, this is a substantial improvement. But it surprises me that even after this change, you're getting a PP performance range of `20.83-35.12 t/s`.

That's a huge gap. I can't reproduce this. And in your case, you were getting the max and min performance multiple times! With 6 threads, this doesn't seem explicable just by OS scheduling quirks.

Perhaps it makes sense to wrap the invocation in something like [Hyperfine](https://github.com/sharkdp/hyperfine), which will keep running the benchmark until the statistics stabilize.