@BarfingLemurs 2024-04-26
@Jeximo I couldn't see an android performance boost. Mind sharing your arguments? If -tb is set manually to greater than 4, eg: 5, the speed dramatically decreases from 22 t/s to 7 t/s, testing tinyllama