@USBhost 2024-04-26
`python koboldcpp.py  --threads 24 --contextsize 65536 --model ../Mixtral-8x22B-Instruct-v0.1/ggml-model-Q8_0.gguf`
I'm using koboldcpp with this PR running on a AMD EPYC 7F72 server with nothing running. Inference is actually 'slower' but so small.... This PR may be good for a system that's not idle?   


With PR:
Processing Prompt (158 / 158 tokens)
Generating (120 / 120 tokens)
CtxLimit: 278/65536, Process:12.77s (80.8ms/T = 12.37T/s), Generate:39.64s (330.3ms/T = 3.03T/s), Total:52.41s (2.29T/s)
Processing Prompt (135 / 135 tokens)
Generating (120 / 120 tokens)
CtxLimit: 278/65536, Process:10.96s (81.2ms/T = 12.32T/s), Generate:39.64s (330.3ms/T = 3.03T/s), Total:50.59s (2.37T/s)
Processing Prompt (132 / 132 tokens)
Generating (120 / 120 tokens)
CtxLimit: 278/65536, Process:10.73s (81.3ms/T = 12.30T/s), Generate:39.67s (330.6ms/T = 3.02T/s), Total:50.40s (2.38T/s)

Without:
Processing Prompt (158 / 158 tokens)
Generating (120 / 120 tokens)
CtxLimit: 278/65536, Process:12.73s (80.6ms/T = 12.41T/s), Generate:38.92s (324.3ms/T = 3.08T/s), Total:51.65s (2.32T/s)
Processing Prompt (135 / 135 tokens)
Generating (120 / 120 tokens)
CtxLimit: 278/65536, Process:10.91s (80.8ms/T = 12.38T/s), Generate:38.92s (324.4ms/T = 3.08T/s), Total:49.83s (2.41T/s)
Processing Prompt (135 / 135 tokens)
Generating (120 / 120 tokens)
CtxLimit: 278/65536, Process:10.92s (80.9ms/T = 12.36T/s), Generate:38.89s (324.1ms/T = 3.09T/s), Total:49.81s (2.41T/s)