@kunnis 2024-04-26
Just a note...  for my testing, I've disabled all BLASs, including Llamafile.