@kunnis 2024-04-27
It was all windows scheduling, for testing, I threw in some basic affinity code  (the n*2 is because I have HT enabled, and I want' it on different physical cores)
```
static void set_numa_thread_affinity(int thread_n) {
    UNUSED(thread_n);
    HANDLE handle = GetCurrentThread();
    SetThreadPriority(handle, THREAD_PRIORITY_ABOVE_NORMAL);
    SetThreadAffinityMask(handle, 1 << (thread_n*2));
}
```
And I'm getting tight timing results on windows now, with a 4% margin of error on PP, and sub 1% on tg.  I also added timing to ``ggml_compute_forward_mul_mat_one_chunk``   But that means that scheduling stalls are the issue.

@slaren If I write something like my code, but without the janky "Just divide by 16...", where it properly partitions it like #1507 does, would that probably get merged in?  (also, thanks for the link to 1507)    If so, I'll cook that up over the weekend.