@kunnis 2024-04-27
I couldn't sleep, so I coded the idea up.