@LostRuins 2024-04-27
One confounding factor to take account for is the behavior of **E-Cores** vs **P-Cores**. On newer systems a thread running on an E-Core can be significantly slower than one running on a P-Core causing a potential bottleneck not present in older systems, or systems with different scheduling behavior. 

The thread scheduling being OS dependent and unpredictable may make comparisons tricky, might be worth disabling efficient cores first to compare the improvement.

Maybe Related:
https://github.com/LostRuins/koboldcpp/issues/447