@kunnis 2024-04-27
I'm just running a Ryzen 7600, so I'm not having the e-core/p-core situation, but it should address that issue as a side effect.  The updated PR divides most MMs of a 7b model into 10000+ chunks, and threads that were previously finishing up to 4000us apart (even with prioritization that I hacked in for testing) are now finishing within about 20us of each other.   And this should be a pretty clean solution that's cross platform.  Instead of assuming all threads will get the same amount of execution time, just divide the larger MM into small blocks, and execute the smaller blocks.