@kunnis 2024-04-27
I removed the global, moving it to shared state, but that required moving a bit of code around to allow that.  Next I'll go through and fixed build errors.  It's just unused variables from some configs, I just need to address them.

Maybe we should move the thread configuration (atomic_store, ggml_thread_create, set_numa_thread_affinity, etc) to it's own file?

I also don't like that I had to pass the ggml_compute_state through ggml_compute_forward, but I don't see a better way of doing it.  I mean it kinda seems right, because everything needs arguments like ith and nth that originate on state.