@slaren 2024-04-30
13900k under WSL2:

| model                          |       size |  backend    |    threads | test       |       master t/s |         PR t/s |    speedup |
| ------------------------------ | ---------: |  ---------- | ---------: | ---------- | ---------------: | -------------: | ---------: |
| llama 7B Q4_K - Medium         |   3.80 GiB |  CPU        |          8 | pp 128     |     47.47 ± 0.84 |   56.79 ± 0.87 |  1.20 |
| llama 7B Q4_K - Medium         |   3.80 GiB |  CPU        |          8 | tg 32      |     16.01 ± 0.22 |   18.00 ± 0.10 |  1.12 |
| llama 7B Q4_K - Medium         |   3.80 GiB |  CPU        |         16 | pp 128     |     42.67 ± 0.27 |   72.97 ± 1.01 |  1.71 |
| llama 7B Q4_K - Medium         |   3.80 GiB |  CPU        |         16 | tg 32      |     14.45 ± 0.05 |   18.83 ± 0.03 |  1.30 |
| llama 7B Q4_K - Medium         |   3.80 GiB |  CPU        |         24 | pp 128     |     60.48 ± 0.37 |   85.12 ± 0.89 |  1.41 |
| llama 7B Q4_K - Medium         |   3.80 GiB |  CPU        |         24 | tg 32      |     16.17 ± 0.04 |   18.72 ± 0.04 |  1.16 |
| llama 7B Q4_K - Medium         |   3.80 GiB |  CPU        |         32 | pp 128     |     68.64 ± 0.58 |   84.54 ± 2.63 |  1.23 |
| llama 7B Q4_K - Medium         |   3.80 GiB |  CPU        |         32 | tg 32      |     14.20 ± 0.15 |   15.52 ± 0.08 |  1.09 |

So it's a very nice speedup for a CPU with e-cores. I didn't expect that it would also improve generation performance with 8 threads (=number of p-cores), I guess the OS scheduler is not doing a very good job here. Perplexity also looks good. I will try under native Windows as well later.

I think it is important to keep the option of using the previous fixed-scheduling type though, I expect will be especially important for numa systems, and it may also help with systems with only one type of CPU core.