@slaren 2024-04-30
13900k with MSVC

| model                          |       size | backend    |    threads | test       |       master t/s |           PR t/s |  speedup |
| ------------------------------ | ---------: | ---------- | ---------: | ---------- | ---------------: | ---------------: | -------: |
| llama 7B Q4_K - Medium         |   3.80 GiB | CPU        |          8 | pp 128     |     42.76 ± 0.46 |     49.61 ± 0.11 |     1.16 |
| llama 7B Q4_K - Medium         |   3.80 GiB | CPU        |          8 | tg 32      |     15.59 ± 0.23 |     17.23 ± 0.07 |     1.11 |
| llama 7B Q4_K - Medium         |   3.80 GiB | CPU        |         16 | pp 128     |     37.72 ± 0.03 |     63.41 ± 0.82 |     1.68 |
| llama 7B Q4_K - Medium         |   3.80 GiB | CPU        |         16 | tg 32      |     11.91 ± 0.04 |     16.37 ± 0.01 |     1.37 |
| llama 7B Q4_K - Medium         |   3.80 GiB | CPU        |         24 | pp 128     |     52.54 ± 0.17 |     72.59 ± 0.29 |     1.38 |
| llama 7B Q4_K - Medium         |   3.80 GiB | CPU        |         24 | tg 32      |     12.51 ± 0.02 |     14.99 ± 0.06 |     1.20 |
| llama 7B Q4_K - Medium         |   3.80 GiB | CPU        |         32 | pp 128     |     33.28 ± 0.56 |     67.71 ± 1.78 |     2.03 |
| llama 7B Q4_K - Medium         |   3.80 GiB | CPU        |         32 | tg 32      |      4.03 ± 0.11 |      6.50 ± 0.34 |     1.61 |

Similar speedup, but worse absolute performance.